# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for the Diploma project.",
        "title": "Diploma Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/homes/{home_id}/polls": {
            "get": {
                "description": "Get all polls in a home",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "poll"
                ],
                "summary": "Get all polls by home ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Home ID",
                        "name": "home_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "description": "Create a new poll in a home",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "poll"
                ],
                "summary": "Create a new poll",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Home ID",
                        "name": "home_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Poll Request",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatePollRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/homes/{home_id}/polls/{poll_id}": {
            "get": {
                "description": "Get poll details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "poll"
                ],
                "summary": "Get poll by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Home ID",
                        "name": "home_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Poll ID",
                        "name": "poll_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "description": "Delete a poll by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "poll"
                ],
                "summary": "Delete poll",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Home ID",
                        "name": "home_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Poll ID",
                        "name": "poll_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/homes/{home_id}/polls/{poll_id}/close": {
            "patch": {
                "description": "Close a poll by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "poll"
                ],
                "summary": "Close poll",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Home ID",
                        "name": "home_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Poll ID",
                        "name": "poll_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/homes/{home_id}/polls/{poll_id}/vote": {
            "post": {
                "description": "Vote in a poll",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "poll"
                ],
                "summary": "Vote in poll",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Home ID",
                        "name": "home_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Poll ID",
                        "name": "poll_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vote Request",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VoteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "models.CreatePollRequest": {
            "type": "object",
            "required": [
                "question",
                "type"
            ],
            "properties": {
                "options": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.OptionRequest"
                    }
                },
                "question": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "single",
                        "multiple"
                    ]
                }
            }
        },
...
        "models.OptionRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.VoteRequest": {
            "type": "object",
            "required": [
                "option_id"
            ],
            "properties": {
                "option_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}